fixed issues

+++ src/repositories/bookRepository.js
@@ -6,12 +6,20 @@ const addBook = async (data) => {
 };
 
 const deleteBook = async (id) => {
+  // Find the book to delete
+  const bookToDelete = await BookModel.findById(id);
+  if (!bookToDelete) {
+    throw new Error("Book does not exist");
+  }
   return await BookModel.findByIdAndDelete(id);
 };
 
 const updateBook = async (id, data) => {
   // Find the book to update
-  const bookToUpdate = BookModel.findById(id);
+  const bookToUpdate = await BookModel.findById(id);
+  if (!bookToUpdate) {
+    throw new Error("Book does not exist");
+  }
   const newBook = new BookModel(Object.assign(bookToUpdate, data));
   return newBook.save();
 };
